Execute command at user prompt startup
The last commit added a user prompt when the user takes an action (e.g.
run a file) when the server is not yet started. The prompt asks the user
whether the server should be started.

While this function was already working, it was not very convenient yet
because the user would have to wait for the server to start up (in the
hidden tmux pane) and then tell the plugin again to run the file. This
commit fixes this problem by adding the command to be executed (e.g. the
current file that the user wants to run) to the MATLAB start up command.