Improve projectroot determination
matlab.vim tries opening MATLAB with the correct working directory. For
this, it uses the projectroot function made by Daan Bakker and checks
furthermore if there is a folder that contains the name matlab in it.

This commit fixes a bug where matlab.vim would try to change the working
directory into a file (e.g. if you have a file called matlab-data.mat in
your root folder).