Add new MATLAB functions to syntax highlighting
The original matlab.vim syntax file was written by Fabrice Guy and last
updated in 2009. This commit brings the highlighting up to the latest
MATLAB release at the time of writing (R2022a).

It also adds a Python script to automatically parse the relevant
reference page of the MATLAB documentation and convert it into a vim
syntax file.