Fix error in block comment highlighting
Matlab block comments are quite peculiar. A block comment starts with a
"%{" and ends with a "%}". If there are any other characters (other than
whitespace) on the same line, the following block is not considered a
valid block comment. Some examples:

Example 1:
%{
This is a comment
%}

    %{
This is also a comment
%}

    %{
Still a comment
    %}

    %{
Comment
    %}

Example 2:
%{
Nested block comment exist too.
        %{
        This is a nested comment.
        %}
This is still a comment
%}

Example 3:
    %{  a
This is not a comment, because of the a at the end.
    %}

Example 4:
%{
Same applies for the closing bracket, because of the a
%} a
This is still a comment
%}

Example 6:
disp('test') %{
This is not a comment
%}

This commit adds the correct behaviour for the last three examples.
Matlab documentation on the issue can be found here:

https://mathworks.com/help/matlab/matlab_prog/comments.html