Improve MATLAB pane handling
The plugin will now automatically detect if the server pane has been
closed between two actions (e.g. MATLAB crashes, user enters quit). If
the pane can not be found, the user is asked whether to restart the
MATLAB server.

Additionally, the pane index is now stored more robustly as a unique ID
including the session and window index. This should make it possible to
use a pane outside the current window or session.