Update MATLAB function definitions
Replace python script with a MATLAB script to find all available
functions. The script can generate two versions of the function list:
1. A list containing all functions, classes, packages and sub-members of
   each package (63k entries)
2. A reduced list not recursing into the packages (14k entries).

The issues with recursing into the packages for option 1 is that each
member will get exported as a syntax word. So for example, matlab
contains a member package called xml. That means that now all mentions
of xml will get highlighted in the source code.

A solution would be only matching matlab.xml in the given example, but I
haven't gotten that to work yet.

The script is limited to the amount of packages installed on the
machine, so I might still not have gotten all possible functions. MATLAB
is huge!